Move evil_genius_labs base entity to separate module
