+++ homeassistant/components/evil_genius_labs/__init__.py
@@ -7,9 +7,7 @@ import pyevilgenius
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import aiohttp_client, device_registry as dr
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
+from homeassistant.helpers import aiohttp_client
 
 from .const import DOMAIN
 from .coordinator import EvilGeniusUpdateCoordinator
@@ -41,23 +39,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-class EvilGeniusEntity(CoordinatorEntity[EvilGeniusUpdateCoordinator]):
-    """Base entity for Evil Genius."""
-
-    _attr_has_entity_name = True
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device info."""
-        info = self.coordinator.info
-        return DeviceInfo(
-            identifiers={(DOMAIN, info["wiFiChipId"])},
-            connections={(dr.CONNECTION_NETWORK_MAC, info["macAddress"])},
-            name=self.coordinator.device_name,
-            model=self.coordinator.product_name,
-            manufacturer="Evil Genius Labs",
-            sw_version=info["coreVersion"].replace("_", "."),
-            configuration_url=self.coordinator.client.url,
-        )
