+++ homeassistant/components/evil_genius_labs/light.py
@@ -11,9 +11,9 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import EvilGeniusEntity
 from .const import DOMAIN
 from .coordinator import EvilGeniusUpdateCoordinator
+from .entity import EvilGeniusEntity
 from .util import update_when_done
 
 HA_NO_EFFECT = "None"
